Handle mixed keys (symbol/string) in params
An issue was reported in #7 that illustrated an error when the params
hash had a mix of string and symbol keys. Ultimately it appears that
grape-entity is creating this inconsistent hash but normalizing the
keys resolves the error.

NOTE: I was not able to reproduce this error in specs because
grape-entity is not included in this project however I did reproduce
and test the fix externally.